+++ packages/hedera-wallet-snap/packages/snap/src/components/signMessage.tsx
@@ -1,0 +1,57 @@
+/*-
+ *
+ * Hedera Wallet Snap
+ *
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import type { SnapComponent } from '@metamask/snaps-sdk/jsx';
+import {
+  Box,
+  CommonUI,
+  Copyable,
+  Heading,
+  Text,
+  type CommonProps,
+} from './common';
+
+type Props = CommonProps & {
+  message: string;
+};
+
+const MetamaskUI: SnapComponent<Props> = ({
+  origin,
+  network,
+  mirrorNodeUrl,
+  message,
+}) => {
+  return (
+    <Box>
+      <CommonUI
+        origin={origin}
+        network={network}
+        mirrorNodeUrl={mirrorNodeUrl}
+      />
+      <Heading>Sign Message</Heading>
+      <Box>
+        <Text>Are you sure you want to sign this message?</Text>
+        <Copyable value={message} />
+      </Box>
+    </Box>
+  );
+};
+
+export const SignMessageUI = MetamaskUI as any;