+++ packages/hedera-wallet-snap/packages/snap/src/facades/hts/CreateTokenFacade.ts
@@ -24,7 +24,7 @@ import { copyable, divider, heading, text } from '@metamask/snaps-sdk';
 import _ from 'lodash';
 import { HederaClientImplFactory } from '../../client/HederaClientImplFactory';
 import { CreateTokenCommand } from '../../commands/hts/CreateTokenCommand';
-import type { TxReceipt } from '../../types/hedera';
+import type { TxRecord } from '../../types/hedera';
 import type { CreateTokenRequestParams } from '../../types/params';
 import type { WalletSnapParams } from '../../types/state';
 import { SnapUtils } from '../../utils/SnapUtils';
@@ -46,7 +46,7 @@ export class CreateTokenFacade {
   public static async createToken(
     walletSnapParams: WalletSnapParams,
     createTokenRequestParams: CreateTokenRequestParams,
-  ): Promise<TxReceipt> {
+  ): Promise<TxRecord> {
     const { origin, state } = walletSnapParams;
 
     const { hederaEvmAddress, hederaAccountId, network, mirrorNodeUrl } =
@@ -76,7 +76,7 @@ export class CreateTokenFacade {
       maxSupply = 0,
     } = createTokenRequestParams;
 
-    let txReceipt = {} as TxReceipt;
+    let txReceipt = {} as TxRecord;
     try {
       const panelToShow = [
         heading('Create a token'),
@@ -244,6 +244,12 @@ export class CreateTokenFacade {
       txReceipt = await command.execute(
         hederaClient.getClient(),
         privateKeyObj,
+      );
+      await SnapUtils.snapCreateDialogAfterTransaction(
+        origin,
+        network,
+        mirrorNodeUrl,
+        txReceipt,
       );
     } catch (error: any) {
       const errMessage = `Error while trying to create a token`;